+++ pytket/binders/circuit/main.cpp
@@ -376,6 +376,10 @@ PYBIND11_MODULE(circuit, m) {
           "\\mathrm{Rz}(\\beta)\\mathrm{Rx}(\\alpha)\\mathrm{Rz}(-"
           "\\beta)` (matrix-multiplication order)")
       .value(
+          "NPhasedX", OpType::NPhasedX,
+          ":math:`(\\alpha, \\beta)` := "
+          "n-qubit gate equivalent to n PhasedX(α, β) gates in parallel.")
+      .value(
           "CnRy", OpType::CnRy,
           ":math:`(\\alpha)` := n-controlled "
           ":math:`\\mathrm{Ry}(\\alpha)` gate.")